pyproject.toml
[project]
name = "langchain-chainlit-agent-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arxiv>=2.4.0",
    "chainlit>=2.0.0",
    "fastmcp>=3.0.2",
    "feedparser>=6.0.12",
    "langchain>=0.3.0",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-openai>=0.3.0",
    "mcp>=1.0.0",
    "python-dotenv>=1.0.0",
]
